Atualização do PR e mudança no componente de Destaque de adoção
