+++ src/app/page.jsx
@@ -1,3 +1,4 @@
+import { DestaqueAdocao } from "@/components/DestaqueAdocao";
 import { SliderOngs } from "@/components/Home/SliderOngs/SliderOngs";
 import Layout from "@/components/Layout";
 
@@ -5,6 +6,7 @@ export default function Home() {
   return (
     <>
       <Layout>
+        <DestaqueAdocao />
         <SliderOngs />
       </Layout>
     </>
